finish drawBmpFromFile refactoring for 24-bit direct

+++ src/MiniGrafx.h
@@ -128,7 +128,7 @@ class MiniGrafx {
   void drawTriangle(uint16_t x1, uint16_t y1, uint16_t x2, uint16_t y2, uint16_t x3, uint16_t y3);
   uint16_t getStringWidth(const char* text, uint16_t length);
   void drawXbm(int16_t x, int16_t y, int16_t width, int16_t height, const char *xbm);
-  void drawBmpFromFile(String filename, uint8_t x, uint16_t y, bool directWrite = false);
+  void drawBmpFromFile(String filename, int16_t x, int16_t y, bool directWrite = false);
   void drawBmpFromPgm(const char *xbm, uint8_t x, uint16_t y);
   void drawPalettedBitmapFromPgm(uint16_t x, uint16_t y, const char *palBmp);
   void drawPalettedBitmapFromFile(uint16_t x, uint16_t y, String fileName);
